// src/app/modules/characters/models/character.model.ts
export interface CharacterI {
  birth_year: any;
  eye_color: string;
  films: any[];
  gender: string;
  hair_color: string;
  height: number;
  homeworld: string;
  mass: number;
  name: string;
  skin_color: string;
  created: Date;
  edited: Date;
  species: any[];
  starships: any[];
  url: string;
  vehicles: any[];
}