// src/app/modules/characters/models/species.model.ts
export interface SpeciesI {
  name: string;
  classification: string;
  designation: string;
  average_height: number;
  skin_colors: string;
  hair_colors: string;
  eye_colors: string;
  average_lifespan: number;
  homeworld: string;
  language: string;
  people: any[];
  films: any[];
  created: Date;
  edited: Date;
  url: string;
}